Refactor test classes and setup for Selenium with TestNG
1. Separated Test Classes: Refactored the main test class SeleniumTestngTest into multiple specific test classes for better organization. Each test case now resides in its own class:

- ABTestingTest
- AddRemoveElementsTest
- CheckboxesTest
- DropdownTest
- FormAuthenticationTest
- DragAndDropTest
- HorizontalSliderTest
2. TestNG Suite Configuration: Updated the TestNG suite XML file to include all the newly separated test classes. This enhances readability and maintainability by keeping related tests in distinct classes.
3. Added README_TEST.md file, which contains instructions for setting up and running tests.
