+++  README_TEST.md
@@ -1,0 +1,161 @@
+# Selenium TestNG (Домашнее задание 13)
+Этот проект содержит автоматические тесты для приложения "The Internet" с использованием Selenium и TestNG. 
+Тесты покрывают различные функциональности, такие как A/B Testing, добавление и удаление элементов, чекбоксы, выпадающий список, аутентификация, перетаскивание элементов и горизонтальный слайдер.
+
+## Начало работы
+### 1. Клонирование репозитория
+
+Для начала клонируйте репозиторий с домашним заданием на локальную машину:
+https://github.com/JAVA-QA-AUTO-UA-RB/2-testng-selenium-css-xpath-locators-IrinaPleshakova.git
+
+### 2. Открытие в IntelliJ IDEA
+Откройте проект в IntelliJ IDEA для работы с кодом и тестами.
+
+#### Требования
+
+Перед запуском тестов убедитесь, что на вашем компьютере установлены:
+
+- Java JDK 11 или выше
+- Apache Maven 3.6 или выше
+- Браузер (Chrome)
+
+## Зависимости
+Проект использует следующие зависимости:
+
+1. Selenium Java - Версия 4.9.1
+2. WebDriverManager - Версия 5.9.2
+3. TestNG - Версия 7.8.0
+Эти зависимости указаны в pom.xml и будут автоматически загружены при сборке проекта с помощью Maven.
+
+## Структура проекта
+```shell
+2-testng-selenium-css-xpath-locators-IrinaPleshakova/
+├── src/
+│   ├── main/
+│   │   └── java/
+│   │       └── example/
+│   │           └── Main.java
+│   └── test/
+│       └── java/
+│           ├── ABTestingTest.java
+│           ├── AddRemoveElementsTest.java
+│           ├── BasicSetupTest.java
+│           ├── CheckboxesTest.java
+│           ├── DragAndDropTest.java
+│           ├── DropdownTest.java
+│           ├── FormAuthenticationTest.java
+│           └── HorizontalSliderTest.java
+├── pom.xml
+├── README_TEST.md
+├── README.md
+└── testng.xml
+```
+
+### Описание файлов:
+ - ABTestingTest.java: Тесты для страницы A/B Testing.
+ - AddRemoveElementsTest.java: Тесты для страницы добавления/удаления элементов.
+ - BasicSetupTest.java: Базовый класс для инициализации браузера.
+ - CheckboxesTest.java: Тесты для страницы с чекбоксами.
+ - DragAndDropTest.java: Тесты для страницы с перетаскиванием элементов.
+ - DropdownTest.java: Тесты для страницы с выпадающим списком.
+ - FormAuthenticationTest.java: Тесты для страницы аутентификации.
+ - HorizontalSliderTest.java: Тесты для страницы с горизонтальным слайдером.
+ - pom.xml: Конфигурационный файл Maven, содержащий все необходимые зависимости.
+ - README_TEST.md: Инструкции по настройке и запуску тестов.
+ - README.md: Инструкции по выполнению домашнего задания.
+
+### Описание тестов: нажми details
+<details>
+
+#### ABTestingTest.java
+Этот класс включает тесты для страницы A/B Testing.
+
+Основная цель теста — проверить, что на странице присутствует заголовок с текстом "A/B Test Control".
+
+- Открытие страницы A/B Testing: Переход по ссылке к странице.
+- Проверка заголовка: Находится элемент с заголовком и проверяется его отображение и правильность текста.
+- Дополнительные проверки: Если элемент не отображается или текст не совпадает, тест завершится с ошибкой.
+
+#### AddRemoveElementsTest.java
+Этот класс тестирует функциональность добавления и удаления элементов на странице.
+
+- Открытие страницы: Переход на страницу с кнопками добавления/удаления элементов.
+- Добавление кнопок "Delete": Тест кликает на кнопку "Add Element" три раза и проверяет наличие трех кнопок "Delete".
+- Проверка видимости: Убедитесь, что каждая кнопка "Delete" видима на странице.
+- Удаление кнопок: Все кнопки "Delete" кликаются для удаления, после чего проверяется, что кнопок больше нет на странице.
+
+#### BasicSetupTest.java
+Этот класс является базовым для инициализации и завершения работы браузера перед выполнением тестов и после них.
+
+- Инициализация WebDriver: Перед выполнением тестов происходит автоматическая загрузка и установка последней версии WebDriver для Chrome.
+- Открытие браузера: Создается новый экземпляр ChromeDriver перед запуском тестов.
+- Закрытие браузера: После выполнения всех тестов браузер закрывается, чтобы завершить сессию.
+
+#### CheckboxesTest.java
+Этот класс тестирует страницу с чекбоксами.
+
+- Открытие страницы: Переход на страницу с чекбоксами.
+- Установка чекбоксов: Если чекбоксы не выбраны, тест кликает по ним, чтобы установить их в выбранное состояние.
+- Проверка состояния: Тест проверяет, что чекбоксы действительно находятся в состоянии "выбрано".
+- Дополнительные проверки: Проверка наличия атрибута "checked" у каждого чекбокса.
+
+#### DragAndDropTest.java
+Класс тестирует функциональность перетаскивания элементов на странице Drag and Drop.
+
+- Открытие страницы: Переход на страницу с элементами для перетаскивания.
+- Перетаскивание элементов: Тест выполняет перетаскивание элемента A на место элемента B с помощью Actions.
+- Проверка результата: Убедитесь, что элементы действительно поменялись местами.
+- Дополнительные проверки: Элементы остаются доступными для взаимодействия после выполнения действия.
+
+#### DropdownTest.java
+Этот класс тестирует страницу с выпадающим списком.
+
+- Открытие страницы: Переход на страницу с выпадающим списком.
+- Выбор опции: Тест выбирает "Option 2" из выпадающего списка и проверяет, что эта опция стала активной.
+- Проверка доступности: Убедитесь, что после выбора опции выпадающий список остается доступным для взаимодействия.
+
+#### FormAuthenticationTest.java
+Этот класс тестирует страницу аутентификации.
+
+- Открытие страницы: Переход на страницу с формой входа.
+- Ввод данных: Ввод корректного логина и пароля для выполнения входа.
+- Проверка успешного входа: Тест проверяет наличие сообщения об успешной авторизации и кнопки "Logout".
+- Выход из системы: Нажатие на кнопку "Logout" и проверка, что пользователь вышел из системы.
+- Дополнительные проверки: Проверка наличия сообщения о выходе и доступности кнопки входа после завершения сеанса.
+
+#### HorizontalSliderTest.java
+Этот класс тестирует функциональность горизонтального слайдера.
+
+- Открытие страницы: Переход на страницу с горизонтальным слайдером.
+- Перемещение слайдера: Тест перемещает слайдер с начального значения 0 до значения 3.
+- Проверка изменения значения: Убедитесь, что слайдер действительно изменил значение на 3.
+- Дополнительные проверки: Проверка видимости и доступности слайдера для дальнейшего взаимодействия.
+</details>
+
+
+
+## Запуск тестов
+
+Для выполнения тестов используйте команду: `mvn clean test`
+
+## Пример вывода в консоли
+Если тесты прошли успешно, вы увидите следующий вывод в консоли:
+
+```shell
+[INFO] -------------------------------------------------------
+[INFO]  T E S T S
+[INFO] -------------------------------------------------------
+[INFO] Running TestSuite
+[INFO] Tests run: 7, Failures: 0, Errors: 0, Skipped: 0, Time elapsed: 6.688 s - in TestSuite
+[INFO] 
+[INFO] Results:
+[INFO] 
+[INFO] Tests run: 7, Failures: 0, Errors: 0, Skipped: 0
+[INFO] 
+[INFO] ------------------------------------------------------------------------
+[INFO] BUILD SUCCESS
+[INFO] ------------------------------------------------------------------------
+[INFO] Total time:  9.488 s
+[INFO] Finished at: 2024-08-24T01:03:03+02:00
+[INFO] ------------------------------------------------------------------------
+
