+++ testng.xml
@@ -1,8 +1,16 @@
 <!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd" >
 <suite name="The Internet App Tests" verbose="1">
-    <test name="[TheInternetApp] Functional Smoke">
-        <classes>
-            <class name="SeleniumTestngTest"/>
-        </classes>
-    </test>
+
+<test name="[TheInternetApp] Functional Smoke">
+    <classes>
+        <class name="ABTestingTest"/>
+        <class name="AddRemoveElementsTest"/>
+        <class name="CheckboxesTest"/>
+        <class name="DropdownTest"/>
+        <class name="FormAuthenticationTest"/>
+        <class name="DragAndDropTest"/>
+        <class name="HorizontalSliderTest"/>
+    </classes>
+</test>
+
 </suite>
